requirements.txt
behavex
selenium
allure-behave